+++ components/Dashboard.jsx
@@ -7,8 +7,9 @@ import CardFarming from './automation/CardFarming';
 import Achievements from './achievements/Achievements';
 import AchievementUnlocker from './automation/AchievementUnlocker';
 import { antiAwayStatus } from '@/utils/utils';
+import FreeGamesList from './gameslist/FreeGamesList';
 
-export default function Dashboard({ userSummary, setUserSummary, setInitUpdate, setUpdateManifest }) {
+export default function Dashboard({ userSummary, setUserSummary, setInitUpdate, setUpdateManifest, showFreeGamesTab, freeGamesList }) {
     const [activePage, setActivePage] = useState('games');
     const [appId, setAppId] = useState(null);
     const [appName, setAppName] = useState(null);
@@ -29,7 +30,7 @@ export default function Dashboard({ userSummary, setUserSummary, setInitUpdate, 
                     <Achievements steamId={userSummary.steamId} appId={appId} appName={appName} setShowAchievements={setShowAchievements} isQuery={isQuery} />
                 ) : activePage === 'games' ? (
                     <React.Fragment>
-                        <SideBar setUserSummary={setUserSummary} activePage={activePage} setActivePage={setActivePage} />
+                        <SideBar setUserSummary={setUserSummary} activePage={activePage} setActivePage={setActivePage} showFreeGamesTab={showFreeGamesTab} />
                         <GamesList
                             steamId={userSummary.steamId}
                             inputValue={inputValue}
@@ -41,9 +42,14 @@ export default function Dashboard({ userSummary, setUserSummary, setInitUpdate, 
                             setShowAchievements={setShowAchievements}
                         />
                     </React.Fragment>
+                ) : activePage === 'freeGames' ? (
+                    <React.Fragment>
+                        <SideBar setUserSummary={setUserSummary} activePage={activePage} setActivePage={setActivePage} showFreeGamesTab={showFreeGamesTab} />
+                        <FreeGamesList freeGamesList={freeGamesList} />
+                    </React.Fragment>
                 ) : activePage === 'settings' ? (
                     <React.Fragment>
-                        <SideBar setUserSummary={setUserSummary} activePage={activePage} setActivePage={setActivePage} />
+                        <SideBar setUserSummary={setUserSummary} activePage={activePage} setActivePage={setActivePage} showFreeGamesTab={showFreeGamesTab} />
                         <Settings setInitUpdate={setInitUpdate} setUpdateManifest={setUpdateManifest} />
                     </React.Fragment>
                 ) : activePage === 'card-farming' ? (