+++ changelog.md
@@ -1,6 +1,7 @@
 [![Downloads](https://img.shields.io/github/downloads/probablyraging/steam-game-idler/1.5.17/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/probablyraging/steam-game-idler/releases/download/1.5.17/Steam.Game.Idler_1.5.17_x64_en-US.msi)
 
 ## Changelog
-- Improved the method for checking for free games
-  - SGI will now check for free games - once on launch, and again every hour while SGI is running
-  - Notifications about free games being available will only be shown once every 24 hours
+- Further improvements of the `free games` features
+  - Now, if there are any free Steam games available, there will be a `gift` icon below the `games list` icon in the sidebar. Clicking on it will display a list of the currently free games
+  - Removed the notification associated with free games as it is no longer needed
+  - Removed the `get notified about free games` option in `settings > general` as it is no longer needed